foodmagic-backend/requirements.txt
FastAPI
uvicorn
python-dotenv
openai
pillow